+++ latex_to_myst/macros.tex
@@ -1,29 +1,13 @@
-\DeclareMathOperator{\sinc}{sinc}
-\DeclareMathOperator{\argmin}{argmin}
-\DeclareMathOperator{\supp}{supp}
-
-\def\expect{{\mathbb  E}}
-\def\Pr{{\mathbb P}}
-\def\qed{\hfill{$\Box$} \\}
-\def\eqdef{\triangleq}
-\def\complex{{\mathbb C}}
-\def\real{{\mathbb  R}}
-\def\nat{{\mathbb  N}}
-\def\integer{{\mathbb Z}}
-
+\newtheorem{axiom}{Axiom}
 \newtheorem{theorem}{Theorem}
-\newtheorem{result}{Result}
-\newtheorem{proposition}{Proposition}
 \newtheorem{lemma}{Lemma}
-\newtheorem{assumption}{Assumption}
-\newtheorem{corollary}{Corollary}
-\newtheorem{definition}{Definition}
-{\theorembodyfont{\upshape}
-\newtheorem{remark}{Remark}
-\newtheorem{example}{Example}}
 \newtheorem{algorithm}{Algorithm}
-
-\newcommand{\trans}{{\mathsf{T}}}
-\newcommand{\herm}{{\mathsf{H}}}
-\newcommand{\hilbert}{{\mathcal H}}
-\newcommand{\expo}[1]{\operatorname{exp}\left(#1\right)}
+\newtheorem{definition}{Definition}
+\newtheorem{remark}{Remark}
+\newtheorem{conjecture}{Conjecture}
+\newtheorem{corollary}{Corollary}
+\newtheorem{criterion}{Criterion}
+\newtheorem{example}{Example}
+\newtheorem{property}{Property}
+\newtheorem{observation}{Observation}
+\newtheorem{proposition}{Proposition}
