+++ tests/sample_files/nested_divs.md
@@ -1,0 +1,25 @@
+<div>
+
+`````{div} block
+
+<div>
+
+````{div} block
+
+<div>
+
+```{div} block
+
+A
+
+```
+
+</div>
+
+````
+
+</div>
+
+`````
+
+</div>
