+++ tests/sample_files/nested_divs.tex
@@ -1,0 +1,7 @@
+\begin{block}
+    \begin{block}
+        \begin{block}
+            A
+        \end{block}
+    \end{block}
+\end{block}
